// types/types.ts
/* --------------------------------- */
/* --------------------------------- */
/* ------------- TYPES ------------- */
/* --------------------------------- */
/* --------------------------------- */
export type ThemeNames = 'amethyst' | 'obsidian';

export interface Settings {
  exposeAll?: boolean;
  allowRemoval?: boolean;
  allowEdit?: boolean;
  showHeader?: boolean;
  viewMode?: 'grid' | 'list';
}

export interface Authentication {
  /** Whether the user is authenticated (has a valid token to connnect with the backend) */
  isAuthenticated: boolean;
  plaintext: {
    /** The user's master key (in raw plaintext) that will be used to encrypt and decrypt items */
    masterKey: string;
  };
}

export interface Database {
  [key: string]: DatabaseItem;
}

export interface DatabaseItem {
  password: string;
  username?: string; // i.e.: email or username
  description: string;
  _id?: string; // Auto generated by mongo db
}